Split skin editor scene library test to own file
Rather than stuffing more random cases to `TestSceneScreenNavigation`.
